update platform
- Updated the platform directory to 2.3.108 version for
the following example projects:

mpfs-spi-master-slave
mpfs-i2c-master-slave
mpfs-pdma-read-write
mpfs-hal-ddr-demo
mpfs-emmc-sd-write-read
mpfs-corepwm-blink
mpfs-timer-example
mpfs-watchdog interrupt

Jira-id: ESSBM-396

+++ driver-examples/mss/mss-i2c/mpfs-i2c-master-slave/src/platform/mpfs_hal/common/nwc/mss_sgmii.c
@@ -1,15 +1,10 @@
 /*******************************************************************************
- * Copyright 2019-2021 Microchip FPGA Embedded Systems Solutions.
+ * Copyright 2019 Microchip FPGA Embedded Systems Solutions.
  *
  * SPDX-License-Identifier: MIT
  *
- * MPFS HAL Embedded Software
- *
- */
-
-/*******************************************************************************
  * @file mss_sgmii.c
- * @author Microchip-FPGA Embedded Systems Solutions
+ * @author Microchip FPGA Embedded Systems Solutions
  * @brief sgmii related functions
  *
  */
@@ -17,7 +12,6 @@
 #include <string.h>
 #include <stdio.h>
 #include "mpfs_hal/mss_hal.h"
-#include "simulation.h"
 
 #ifdef  MPFS_HAL_HW_CONFIG
 
@@ -451,7 +445,6 @@ void ddr_pvt_calibration(void)
      * Wait for IOEN from power detectors DDR and SGMII - IO enable signal from
      * System Control powers on
      *
-     * From DDR phy SAC spec:
      *      MSS processor releases dce bus to send RPC bits to IO buffer,
      *      setting each to it's programmed mode and then asserts
      *      ioen high at end of this state.
@@ -640,15 +633,15 @@ static void set_early_late_thresholds(uint8_t n_late_threshold, uint8_t p_early_
      * Set the N eye width value
      * bits 31:29 for CH1, bits 28:26  for CH0 in spare control (N eye width value)
      */
-    n_eye_values = (n_late_threshold << SHIFT_TO_CH0_N_EYE_VALUE);
-    n_eye_values |= (n_late_threshold << SHIFT_TO_CH1_N_EYE_VALUE);
+    n_eye_values = (uint32_t)(n_late_threshold << SHIFT_TO_CH0_N_EYE_VALUE);
+    n_eye_values |= (uint32_t)(n_late_threshold << SHIFT_TO_CH1_N_EYE_VALUE);
 
     CFG_DDR_SGMII_PHY->SPARE_CNTL.SPARE_CNTL    = (LIBERO_SETTING_SPARE_CNTL & N_EYE_MASK) | n_eye_values;
 
     /*
      * Set P values
      */
-    p_eye_value = (p_early_threshold << SHIFT_TO_REG_RX0_EYEWIDTH);
+    p_eye_value = (uint32_t)(p_early_threshold << SHIFT_TO_REG_RX0_EYEWIDTH);
     CFG_DDR_SGMII_PHY->CH0_CNTL.CH0_CNTL        = ((LIBERO_SETTING_CH0_CNTL & REG_RX0_EYEWIDTH_P_MASK) | p_eye_value) | REG_RX0_EN_FLAG_N;
     CFG_DDR_SGMII_PHY->CH1_CNTL.CH1_CNTL        = ((LIBERO_SETTING_CH1_CNTL & REG_RX0_EYEWIDTH_P_MASK) | p_eye_value) | REG_RX1_EN_FLAG_N;
